refactor: Introduce separate persistence/database classes
As a first step of separating the database module, create adapters
for all domain models around the main database class.
