+++ src/file_search/TvShowFileSearcher.cpp
@@ -2,6 +2,7 @@
 
 #include "data/tv_show/TvShow.h"
 #include "data/tv_show/TvShowEpisode.h"
+#include "database/TvShowPersistence.h"
 #include "globals/Helper.h"
 #include "globals/Manager.h"
 #include "globals/MessageIds.h"
@@ -55,7 +56,9 @@ void TvShowFileSearcher::reload(bool force)
 
     emit searchStarted(tr("Loading TV Shows..."));
     int episodeCounter = 0;
-    const int episodeSum = database().episodeCount();
+
+    mediaelch::TvShowPersistence persistence{*Manager::instance()->database()};
+    const int episodeSum = persistence.episodeCount();
 
     QVector<TvShow*> dbShows = getShowsFromDatabase(force);
     setupShows(files, episodeCounter, episodeSum);
@@ -81,7 +84,8 @@ TvShowEpisode* TvShowFileSearcher::loadEpisodeData(TvShowEpisode* episode)
 
 void TvShowFileSearcher::reloadEpisodes(const mediaelch::DirectoryPath& showDir)
 {
-    database().clearTvShowInDirectory(showDir);
+    mediaelch::TvShowPersistence persistence{*Manager::instance()->database()};
+    persistence.clearTvShowInDirectory(showDir);
     emit searchStarted(tr("Searching for Episodes..."));
 
     // remove old show object
@@ -119,7 +123,7 @@ void TvShowFileSearcher::reloadEpisodes(const mediaelch::DirectoryPath& showDir)
     scanTvShowDir(path, showDir, contents);
     auto* show = new TvShow(showDir, this);
     show->loadData(Manager::instance()->mediaCenterInterfaceTvShow());
-    database().add(show, path);
+    persistence.add(show, path);
 
     emit searchStarted(tr("Loading Episodes..."));
     emit currentDir(show->title());
@@ -145,7 +149,7 @@ void TvShowFileSearcher::reloadEpisodes(const mediaelch::DirectoryPath& showDir)
     QtConcurrent::blockingMapped(episodes, TvShowFileSearcher::reloadEpisodeData);
 
     for (TvShowEpisode* episode : episodes) {
-        database().add(episode, path, show->databaseId());
+        persistence.add(episode, path, show->databaseId());
         show->addEpisode(episode);
         emit progress(++episodeCounter, episodeSum, m_progressMessageId);
         QApplication::processEvents();
@@ -442,25 +446,21 @@ QVector<EpisodeNumber> TvShowFileSearcher::getEpisodeNumbers(QStringList files)
     return episodes;
 }
 
-Database& TvShowFileSearcher::database()
-{
-    return *Manager::instance()->database();
-}
-
 void TvShowFileSearcher::clearOldTvShows(bool forceClear)
 {
     // clear gui
     Manager::instance()->tvShowModel()->clear();
 
+    mediaelch::TvShowPersistence persistence{*Manager::instance()->database()};
     if (forceClear) {
         // Simply delete all shows
-        database().clearAllTvShows();
+        persistence.clearAllTvShows();
 
     } else {
         // Otherwise, only clear disabled directories and those with autoReload.
         for (const mediaelch::MediaDirectory& dir : asConst(m_directories)) {
             if (dir.autoReload || dir.disabled) {
-                database().clearTvShowsInDirectory(mediaelch::DirectoryPath(dir.path));
+                persistence.clearTvShowsInDirectory(mediaelch::DirectoryPath(dir.path));
             }
         }
     }
@@ -468,6 +468,7 @@ void TvShowFileSearcher::clearOldTvShows(bool forceClear)
 
 void TvShowFileSearcher::setupShowsFromDatabase(QVector<TvShow*>& dbShows, int episodeCounter, int episodeSum)
 {
+    mediaelch::TvShowPersistence persistence{*Manager::instance()->database()};
     for (TvShow* show : dbShows) {
         if (m_aborted) {
             return;
@@ -475,7 +476,7 @@ void TvShowFileSearcher::setupShowsFromDatabase(QVector<TvShow*>& dbShows, int e
 
         show->loadData(Manager::instance()->mediaCenterInterfaceTvShow(), false);
 
-        QVector<TvShowEpisode*> episodes = database().episodes(show->databaseId());
+        QVector<TvShowEpisode*> episodes = persistence.episodes(show->databaseId());
         QtConcurrent::blockingMapped(episodes, TvShowFileSearcher::loadEpisodeData);
         for (TvShowEpisode* episode : episodes) {
             if (episode == nullptr) {
@@ -495,6 +496,8 @@ void TvShowFileSearcher::setupShowsFromDatabase(QVector<TvShow*>& dbShows, int e
 
 void TvShowFileSearcher::setupShows(QMap<QString, QVector<QStringList>>& contents, int& episodeCounter, int episodeSum)
 {
+    mediaelch::TvShowPersistence persistence{*Manager::instance()->database()};
+
     QMapIterator<QString, QVector<QStringList>> it(contents);
     while (it.hasNext()) {
         it.next();
@@ -529,9 +532,9 @@ void TvShowFileSearcher::setupShows(QMap<QString, QVector<QStringList>>& content
         auto* show = new TvShow(mediaelch::DirectoryPath(it.key()), this);
         show->loadData(Manager::instance()->mediaCenterInterfaceTvShow());
         emit currentDir(show->title());
-        database().add(show, path);
+        persistence.add(show, path);
 
-        database().transaction();
+        Manager::instance()->database()->db().transaction();
         QVector<TvShowEpisode*> episodes;
 
         // Setup episodes list
@@ -551,12 +554,12 @@ void TvShowFileSearcher::setupShows(QMap<QString, QVector<QStringList>>& content
 
         // Add episodes to model
         for (TvShowEpisode* episode : asConst(episodes)) {
-            database().add(episode, path, show->databaseId());
+            persistence.add(episode, path, show->databaseId());
             show->addEpisode(episode);
             emit progress(++episodeCounter, episodeSum, m_progressMessageId);
         }
 
-        database().commit();
+        Manager::instance()->database()->db().commit();
         Manager::instance()->tvShowModel()->appendShow(show);
     }
 
@@ -565,6 +568,7 @@ void TvShowFileSearcher::setupShows(QMap<QString, QVector<QStringList>>& content
 
 QMap<QString, QVector<QStringList>> TvShowFileSearcher::readTvShowContent(bool forceReload)
 {
+    mediaelch::TvShowPersistence persistence{*Manager::instance()->database()};
     QMap<QString, QVector<QStringList>> contents;
     for (const mediaelch::MediaDirectory& dir : asConst(m_directories)) {
         if (m_aborted) {
@@ -581,7 +585,7 @@ QMap<QString, QVector<QStringList>> TvShowFileSearcher::readTvShowContent(bool f
         // TODO: Check if necessary?
         // If there are no shows in the database for the directory, reload
         // all shows regardless of forceReload.
-        const int showsFromDatabase = database().showCount(mediaelch::DirectoryPath(dir.path));
+        const int showsFromDatabase = persistence.showCount(mediaelch::DirectoryPath(dir.path));
         if (showsFromDatabase == 0) {
             getTvShows(mediaelch::DirectoryPath(dir.path), contents);
             continue;
@@ -597,6 +601,7 @@ QVector<TvShow*> TvShowFileSearcher::getShowsFromDatabase(bool forceReload)
         return {};
     }
 
+    mediaelch::TvShowPersistence persistence{*Manager::instance()->database()};
     QVector<TvShow*> dbShows;
     for (const mediaelch::MediaDirectory& dir : asConst(m_directories)) {
         if (dir.autoReload) { // Those directories are not read from database.
@@ -605,7 +610,7 @@ QVector<TvShow*> TvShowFileSearcher::getShowsFromDatabase(bool forceReload)
         if (dir.disabled) {
             continue;
         }
-        QVector<TvShow*> showsFromDatabase = database().showsInDirectory(mediaelch::DirectoryPath(dir.path));
+        QVector<TvShow*> showsFromDatabase = persistence.showsInDirectory(mediaelch::DirectoryPath(dir.path));
         if (!showsFromDatabase.isEmpty()) {
             dbShows.append(showsFromDatabase);
         }
