+++ src/network/NetworkReplyWatcher.h
@@ -8,7 +8,7 @@
 /// \brief The NetworkReplyWatcher class takes a QNetworkReply* and watches it.
 /// A timeout is set which aborts the download if no response was received after N seconds.
 ///
-/// \par Example
+/// \details Example
 /// \code{cpp}
 ///    QNetworkReply* reply = network.get(request);
 ///    new NetworkReplyWatcher(this, reply); // will delete itself when the reply is deleted
